+++ Assets/Scripts/Globales/Variables.cs
@@ -6,7 +6,7 @@ public class Variables {
     public static float volumenMusica = 100f;
     public static float Ghealth = 100f;
     //private int _GvidaActual = 100;
-    public static int Gexp = 0;
+    public static float Gexp = 0;
     //private int Glevel = 1;
     public static float Gdamage = 10f;
     public static int Gstat = 0;
